requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3

httpx==0.27.2

chromadb==0.5.23

pymupdf==1.24.14
pillow==10.4.0

numpy==2.1.3
sentence-transformers==3.3.1

python-multipart==0.0.17
